+++ packages/esnext/index.js
@@ -2,7 +2,7 @@
 
 module.exports = {
     // - Parseur
-    parser: 'babel-eslint',
+    parser: '@babel/eslint-parser',
     parserOptions: {
         sourceType: 'module',
         ecmaFeatures: {
@@ -25,23 +25,15 @@ module.exports = {
     },
 
     // - Plugins
-    plugins: ['babel'],
+    plugins: ['@babel'],
 
     // - Règles
     extends: '@pulsanova/base',
     rules: {
         // (Prise en charge Babel, voir parent)
-        // @see https://eslint.org/docs/rules/camelcase
-        'camelcase': ['off'],
-        'babel/camelcase': ['error', {
-            ignoreDestructuring: false,
-            properties: 'never',
-        }],
-
-        // (Prise en charge Babel, voir parent)
         // @see https://eslint.org/docs/rules/new-cap
         'new-cap': ['off'],
-        'babel/new-cap': ['error', {
+        '@babel/new-cap': ['error', {
             newIsCap: true,
             newIsCapExceptions: [],
             capIsNew: false,
@@ -62,42 +54,32 @@ module.exports = {
         // (Prise en charge Babel, voir parent)
         // @see https://eslint.org/docs/rules/object-curly-spacing
         'object-curly-spacing': ['off'],
-        'babel/object-curly-spacing': ['error', 'always'],
+        '@babel/object-curly-spacing': ['error', 'always'],
 
         // (Prise en charge Babel, voir parent)
         // @see https://eslint.org/docs/rules/semi
         'semi': ['off'],
-        'babel/semi': ['error', 'always'],
+        '@babel/semi': ['error', 'always'],
 
         // (Prise en charge Babel, voir parent)
         // @see https://eslint.org/docs/rules/no-unused-expressions
         'no-unused-expressions': ['off'],
-        'babel/no-unused-expressions': ['error', {
+        '@babel/no-unused-expressions': ['error', {
             allowShortCircuit: false,
             allowTaggedTemplates: false,
             allowTernary: true,
         }],
 
-        // (Prise en charge Babel, voir parent)
-        // @see https://eslint.org/docs/rules/quotes
-        'quotes': ['off'],
-        'babel/quotes': ['error', 'single', { allowTemplateLiterals: true }],
-
         // - Le mode strict ne doit jamais être déclaré, babel s'en charge.
         // @see https://eslint.org/docs/rules/strict
         'strict': ['error', 'never'],
-
-        // (Prise en charge Babel, voir parent)
-        // @see https://eslint.org/docs/rules/valid-typeof
-        'valid-typeof': ['off'],
-        'babel/valid-typeof': ['error', { requireStringLiterals: true }],
 
         //
         // - Règles désactivées.
         //
 
         // @see https://eslint.org/docs/rules/no-invalid-this
-        'babel/no-invalid-this': ['off'],
+        '@babel/no-invalid-this': ['off'],
     },
 
     // - Overrides