+++ packages/vue/index.js
@@ -30,6 +30,7 @@ module.exports = {
             node: {
                 extensions: [
                     '.mjs',
+                    '.cjs',
                     '.js',
                     '.jsx',
                     '.ts',