+++ packages/esnext/index.js
@@ -1,5 +1,9 @@
 'use strict';
 
+// - Corrige la prise en charge des plugins chargés dans les configs. partagées.
+// @see https://github.com/eslint/eslint/issues/3458
+require('@rushstack/eslint-patch/modern-module-resolution');
+
 module.exports = {
     // - Parseur
     parser: '@babel/eslint-parser',