+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   Vous devez donc spécifier `parserOptions: { babelOptions: { configFile: "babel.config.js" } }`
   en remplacement la valeur de `configFile` par le chemin vers votre configiration Babel dans votre 
   configuration ESLint pour l'utiliser (sinon le parsing ne sera pas adapté à votre configuration Babel).
+- Ajoute la prise en charge de l'extension `.cjs`.
 - Met à jour les dépendances (et notamment ESLint en version 8.x).
 - Nouvelles règles:
   - Pas de propriétés privées non utilisées dans les classes.