+++ CHANGELOG.md
@@ -17,7 +17,14 @@
   - (TypeScript) Empêche l'utilisation de l'opérateur `void` lorsque c'est inutile / déjà le cas.
   - (TypeScript) Vérifie qu'une assertion non-null n'est pas utilisé avec un opérateur null coalescent (e.g. `foo! ?? 'bar'`).
   - (React) Interdit l'utilisation de namespaces dans le JSX (e.g. `<ns:Comp />`) car non supportés.
-  - (React) Empêche l'utilisation de valeurs invalides pour certains attributs HTML (`rel` uniquement pour le moment).
+  - (React / Vue) Empêche l'utilisation de valeurs invalides pour certains attributs HTML (`rel` uniquement pour le moment).
+  - (Vue) Interdit l'utilisation de `v-text` au profit des `children`.
+  - (Vue) Interdit l'utilisation d'attributs invalides sur les elements `<template>`.
+  - (Vue) Interdit l'utilisation de propriétés non définies dans les components.
+  - (Vue) Empêche l'utilisation de propriétés computed dans `data() {}` vu que celles-ci ne sont pas encore initialisées.
+  - (Vue) Interdit l'usage de l'attribut `tag` (déprécié en Vue 3.x) dans les `<RouterLink>` (ou `<router-link>`).
+  - (Vue) Contrôle la gestion des nouvelles lignes dans la déclation des blocks Vue (`<template>`, `<script>`.
+  - (Vue) Empêche l'utilsation de chaînes quotées inutilement (e.g. `{{ 'Hello' }}`).
 
 ## 1.3.2 (2021-09-18)
 - Corrige le parsing de TypeScript dans la configuration Vue.